LayoutTests/fast/js/dfg-constant-fold-uncaptured-variable-that-is-later-captured-expected.txt

Tests that constant folding an access to an uncaptured variable that is captured later in the same basic block doesn't lead to assertion failures.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS foo(true, 5)[0] is 462
PASS successfullyParsed is true

TEST COMPLETE
